create duckdb file reader class
